Split off natives
